meHungry_project/requirements.txt
decorator==3.4.0
dj-database-url==0.3.0
Django==1.7.3
django-ajax-selects==1.3.5
django-carton==1.2.0
django-forms-bootstrap==3.0.1
django-jsonfield==0.9.13
django-stripe-payments==2.0.0
gunicorn==19.2.1
httplib2==0.9
MySQL-python==1.2.5
oauthlib==0.7.2
psycopg2==2.5.4
py-mysql2pgsql==0.1.5
PyJWT==0.4.1
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.10
PyYAML==3.11
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
stripe==1.20.1
termcolor==1.1.0
